+++ src/Cwaff-Guns/HandCannon.cs
@@ -92,6 +92,7 @@ public class SlappProjectile : MonoBehaviour
         if (!this._slapVictim)
             return;
 
+        PlayerController owner = this._projectile.Owner as PlayerController;
         Vector2 victimPos = this._slapVictim.CenterPosition;
         Vector2 impactPos = this._vfx.GetComponent<tk2dSprite>().WorldCenter;
         foreach (AIActor enemy in StaticReferenceManager.AllEnemies)
@@ -105,7 +106,7 @@ public class SlappProjectile : MonoBehaviour
 
             hh.ApplyDamage(this._slapDamage, Vector2.zero, "SLAPPP", CoreDamageTypes.None, DamageCategory.Collision, true);
             if (!hh.IsBoss && !hh.IsSubboss && enemy.knockbackDoer is KnockbackDoer kb)
-                kb.ApplyKnockback(this._slapAngle, _SLAPPP_FORCE);
+                kb.ApplyKnockback(this._slapAngle, _SLAPPP_FORCE * (owner ? owner.KnockbackMult() : 1f));
             if (enemy.behaviorSpeculator && !enemy.behaviorSpeculator.ImmuneToStun)
                 enemy.behaviorSpeculator.Stun(this._isMastered ? _CLAPPP_STUN : _SLAPPP_STUN);
             GameObject go = SpawnManager.SpawnVFX(HandCannon._ClapppShockwave, victimPos, Quaternion.identity, ignoresPools: true);