+++ src/Cwaff-Guns/Pincushion.cs
@@ -4,7 +4,7 @@ public class Pincushion : CwaffGun
 {
     public static string ItemName         = "Pincushion";
     public static string ShortDescription = "Needless Violence";
-    public static string LongDescription  = "Fires an extraordinarily fast barrage of extremely weak pins that lose accuracy as the clip is depleted. Pins cannot damage anything other than enemies, and are completely blocked by barrels, crates, etc.";
+    public static string LongDescription  = "Fires an extraordinarily fast barrage of extremely weak pins that lose accuracy as the clip is depleted. Pins deal fixed damage, cannot damage anything other than enemies, and are completely blocked by barrels, crates, etc.";
     public static string Lore             = "Even when shot out of a gun, a pin won't be particularly harmful to most of the Gungeon's inhabitants. Two pins won't fare much better, nor three, nor ten. But, by firing thousands of pins per second, one of them will in all likelihood eventually lodge itself in a weak point or some other location extremely compromising to its target. Needless to say, the probabilistic nature of this gun threads the line of practicality, but its high drum capacity doesn't leave it without merits in the right situation.";
 
     internal const int   _SIMULTANEOUS_BULLETS = 4;